+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Unreleased
+## v1.1.1 (2022-09-16)
 
 * Fixed `--pass`, `--restart-if`, and `--restart-if-not` not allowing a leading negative number.
 * Fixed `--pass`, `--restart-if`, and `--restart-if-not` not requiring a value.